refactor: remove sample user data and update import paths
