--- sample_data/users.ts
@@ -1,192 +1,0 @@
-export const sampleUsers = [
-  {
-    "uid": "google_100001",
-    "name": "Alice Johnson",
-    "avatarUrl": "https://example.com/avatar1.jpg",
-    "phone": "+1 (555) 111-2233",
-    "email": "alice.johnson@example.com",
-    "address": {
-      "street": "123 Oak St",
-      "zip": 12345,
-      "city": "Springfield",
-      "state": "IL",
-      "private": false,
-      "coordinates": {
-        "lat": 39.7817,
-        "long": -89.6501
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g02", "5f8a7b2b9d3f2c1a3c5e7g03"]
-  },
-  {
-    "uid": "google_100002",
-    "name": "Bob Smith",
-    "avatarUrl": "https://example.com/avatar2.jpg",
-    "phone": "",
-    "email": "bob.smith@example.com",
-    "address": {
-      "street": "",
-      "zip": 0,
-      "city": "",
-      "state": "",
-      "private": false,
-      "coordinates": {
-        "lat": 0,
-        "long": 0
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g01", "5f8a7b2b9d3f2c1a3c5e7g03"]
-  },
-  {
-    "uid": "google_100003",
-    "name": "Charlie Davis",
-    "avatarUrl": "https://example.com/avatar3.jpg",
-    "phone": "+1 (555) 333-4444",
-    "email": "charlie.davis@example.com",
-    "address": {
-      "street": "456 Elm Ave",
-      "zip": 67890,
-      "city": "Rivertown",
-      "state": "CA",
-      "private": true,
-      "coordinates": {
-        "lat": 37.7749,
-        "long": -122.4194
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g01", "5f8a7b2b9d3f2c1a3c5e7g02", "5f8a7b2b9d3f2c1a3c5e7g04"]
-  },
-  {
-    "uid": "google_100004",
-    "name": "Diana Brown",
-    "avatarUrl": "https://example.com/avatar4.jpg",
-    "phone": "+1 (555) 555-6666",
-    "email": "diana.brown@example.com",
-    "address": {
-      "street": "789 Maple Rd",
-      "zip": 54321,
-      "city": "Hillsville",
-      "state": "NY",
-      "private": false,
-      "coordinates": {
-        "lat": 40.7128,
-        "long": -74.0060
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g03", "5f8a7b2b9d3f2c1a3c5e7g05"]
-  },
-  {
-    "uid": "google_100005",
-    "name": "Ethan Green",
-    "avatarUrl": "https://example.com/avatar5.jpg",
-    "phone": "",
-    "email": "ethan.green@example.com",
-    "address": {
-      "street": "101 Pine St",
-      "zip": 98765,
-      "city": "Mountainview",
-      "state": "CO",
-      "private": true,
-      "coordinates": {
-        "lat": 39.7392,
-        "long": -104.9903
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g04", "5f8a7b2b9d3f2c1a3c5e7g06"]
-  },
-  {
-    "uid": "google_100006",
-    "name": "Fiona White",
-    "avatarUrl": "https://example.com/avatar6.jpg",
-    "phone": "+1 (555) 777-8888",
-    "email": "fiona.white@example.com",
-    "address": {
-      "street": "",
-      "zip": 0,
-      "city": "",
-      "state": "",
-      "private": false,
-      "coordinates": {
-        "lat": 0,
-        "long": 0
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g05", "5f8a7b2b9d3f2c1a3c5e7g07"]
-  },
-  {
-    "uid": "google_100007",
-    "name": "George Taylor",
-    "avatarUrl": "https://example.com/avatar7.jpg",
-    "phone": "+1 (555) 999-0000",
-    "email": "george.taylor@example.com",
-    "address": {
-      "street": "202 Cedar Ln",
-      "zip": 13579,
-      "city": "Lakeside",
-      "state": "FL",
-      "private": false,
-      "coordinates": {
-        "lat": 27.9944,
-        "long": -82.4324
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g06", "5f8a7b2b9d3f2c1a3c5e7g08"]
-  },
-  {
-    "uid": "google_100008",
-    "name": "Hannah Lee",
-    "avatarUrl": "https://example.com/avatar8.jpg",
-    "phone": "+1 (555) 123-4567",
-    "email": "hannah.lee@example.com",
-    "address": {
-      "street": "303 Birch Ave",
-      "zip": 24680,
-      "city": "Sunnydale",
-      "state": "TX",
-      "private": true,
-      "coordinates": {
-        "lat": 29.7604,
-        "long": -95.3698
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g07", "5f8a7b2b9d3f2c1a3c5e7g09"]
-  },
-  {
-    "uid": "google_100009",
-    "name": "Ian Clark",
-    "avatarUrl": "https://example.com/avatar9.jpg",
-    "phone": "",
-    "email": "ian.clark@example.com",
-    "address": {
-      "street": "404 Willow Dr",
-      "zip": 35791,
-      "city": "Meadowville",
-      "state": "WA",
-      "private": false,
-      "coordinates": {
-        "lat": 47.6062,
-        "long": -122.3321
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g08", "5f8a7b2b9d3f2c1a3c5e7g10"]
-  },
-  {
-    "uid": "google_100010",
-    "name": "Julia Scott",
-    "avatarUrl": "https://example.com/avatar10.jpg",
-    "phone": "+1 (555) 987-6543",
-    "email": "julia.scott@example.com",
-    "address": {
-      "street": "",
-      "zip": 0,
-      "city": "",
-      "state": "",
-      "private": false,
-      "coordinates": {
-        "lat": 0,
-        "long": 0
-      }
-    },
-    "friends": ["5f8a7b2b9d3f2c1a3c5e7g09", "5f8a7b2b9d3f2c1a3c5e7g01"]
-  }
-]
