+++ src/Email.ts
